Ignore target files
The resulting deb files inside the "target" directory are not ignored by Git,
therefore the output of "git status" is not clean since there are untracked
changes.

Everything under the "target" directory should be ignored, hence adding a
".gitignore" file for it. Also, the ".gitkeep" file is no longer necessary since
the "target" directory will not be empty thanks to the ".gitignore" file we have
added.